+++ app/src/main/java/com/pnj/saku_planner/kakeibo/presentation/screens/report/SummaryScreen.kt
@@ -225,7 +225,7 @@ fun SummaryScreen(
                     }
 
                     state.summaryData.forEachIndexed { index, data ->
-                        val pct = data.amount / total * 100.0
+                        val pct = (data.amount / total.toDouble()) * 100.0
                         val formatted = String.format(Locale.getDefault(), "%.2f", pct)
 
                         Row(
